feat: migrate PWA push notifications functionality and remove deprecated push notification services
