+++ core/src/db/schema/index.ts
@@ -1,5 +1,5 @@
 export * from "@/zap/ai/db/schema";
 export * from "@/zap/auth/db/schema";
+export * from "@/zap/pwa/db/schema";
 export * from "@/zap-old/db/schema/feedbacks.sql";
-export * from "@/zap-old/db/schema/notifications.sql";
 export * from "@/zap-old/db/schema/waitlist.sql";
