+++ core/src/rpc/router.ts
@@ -3,9 +3,9 @@ import "server-only";
 import { ai } from "@/zap/ai/rpc/procedures";
 import { auth } from "@/zap/auth/rpc/procedures";
 import { mails } from "@/zap/mails/rpc/procedures";
+import { pwa } from "@/zap/pwa/rpc/procedures";
 import { example } from "@/zap-old/rpc/procedures/example.rpc";
 import { feedbacks } from "@/zap-old/rpc/procedures/feedbacks.rpc";
-import { pushNotifications } from "@/zap-old/rpc/procedures/push-notifications.rpc";
 import { users } from "@/zap-old/rpc/procedures/users.rpc";
 import { waitlist } from "@/zap-old/rpc/procedures/waitlist.rpc";
 
@@ -15,7 +15,7 @@ export const router = {
   example,
   feedbacks,
   mails,
-  pushNotifications,
+  pwa,
   users,
   waitlist,
 };
