--- core/src/zap-old/rpc/procedures/push-notifications.rpc.ts
@@ -1,22 +1,0 @@
-import "server-only";
-
-import { authMiddleware, base } from "@/rpc/middlewares";
-import { withRpcHandler } from "@/zap/lib/api/handlers";
-import { InputSubscribeUserSchema } from "@/zap/schemas/push-notifications.schema";
-import { subscribeUserService } from "@/zap/services/push-notifications/subscribe-user.service";
-import { unsubscribeUserService } from "@/zap/services/push-notifications/unsubscribe-user.service";
-
-const subscribeUser = base.input(InputSubscribeUserSchema).handler(
-  withRpcHandler(async ({ input }) => {
-    return await subscribeUserService({ ...input });
-  }),
-);
-
-const unsubscribeUser = base
-  .use(authMiddleware)
-  .handler(withRpcHandler(unsubscribeUserService));
-
-export const pushNotifications = {
-  subscribeUser,
-  unsubscribeUser,
-};
