+++ core/src/zap/pwa/db/queries/index.ts
@@ -5,8 +5,8 @@ import { eq, sql } from "drizzle-orm";
 import { db } from "@/db";
 import { pushNotifications } from "@/db/schema";
 
-export const getPushNotificationsByUserQuery = db.query.pushNotifications
+export const getPushNotificationsByUserIdQuery = db.query.pushNotifications
   .findFirst({
     where: eq(pushNotifications.userId, sql.placeholder("userId")),
   })
-  .prepare("getPushNotificationsByUser");
+  .prepare("getPushNotificationsByUserId");
